Update Dockerfile to reduce shipping size with staged build.
In the inital stage, use the `alpine` version of node, mainly to align
with the final shipping stage base layer. Build the dependencies in the
initial stage to avail it in the final stage.

In the shipping or final stage, use `alpine`. There is room for further
improvement here, arguably but using an `alpine` is simple with minimal
hinderances. Install the runtime in the stage and copy the source and
module dependencies from the initial stage.

This approach seem to have significantly helped bringing the size down
by a few factors.