(bug) Profile selection was broken
When selecting the "Profiles" we see all profiles organized by tier.
Here we have both ClusterProfile (which are cluster wide resources)
and Profiles (which are namespaced resources).
While selecting a ClusterProfile was correctly loading a page showing
the ClusterProfile dependencies and dependents, selecting a Profile
was not loading.

The issue is a Profile is identified by namespace and name (while a
ClusterProfile is identified just by name).

This PR fixes that (verified it locally).
This PR also makes sure we consistently use "profiles" when generically
referencing to both ClusterProfile and Profile.